Review feedback: pass generic actions to tooltip (and chart)
